WebApi/Application/DirectorOperations/Commands/CreateDirector/CreateDirectorCommand.cs
using System;
using System.Linq;
using AutoMapper;
using WebApi.DBOperations;
using WebApi.Entities;

namespace WebApi.Application.DirectorOperations.Commands.CreateDirector
{
    public class CreateDirectorCommand
    {
        public CreateDirectorModel Model { get; set; }
        private readonly IMovieStoreDbContext _dbContext;
        private readonly IMapper _mapper;

        public CreateDirectorCommand(IMovieStoreDbContext dbContext, IMapper mapper)
        {
            _dbContext = dbContext;
            _mapper = mapper;
        }

        public void Handle()
        {
            var director = _dbContext.Directors.SingleOrDefault(
                        director => director.FirstName.ToLower() == Model.FirstName.ToLower() 
                        && director.LastName.ToLower() == Model.LastName.ToLower());
            if (director is not null)
            throw new InvalidOperationException("Yönetmen zaten mevcut.");

            director = _mapper.Map<Director>(Model);
            _dbContext.Directors.Add(director);
            _dbContext.SaveChanges();
        }
    }

    public class CreateDirectorModel
    {
        public string FirstName { get; set; }
        public string LastName { get; set; }
    }
}